Prepare test case for `typealias` implementation
Third step is to prepare first attempt to arm a macro into `typealias` code generation. Tere are two cases that will appear later. Firstly macro expands at the end of the code so this expanded source will be automatically wrong. Second problem that should be handed is to catch the latest response body type.